Implement IconEditor in a better way.
