+++ app/src/common/docs/componentEditor/view/PropSamplesCreationContext.tsx
@@ -1,7 +1,9 @@
 import * as React from 'react';
 import { FlexRow, Panel } from '@epam/uui';
+import { Icon } from '@epam/uui-core';
 import { IPropSamplesCreationContext } from '@epam/uui-docs';
 import { svc } from '../../../../services';
+import { getIconList } from '../../../../documents/iconListHelpers';
 
 export class PropSamplesCreationContext implements IPropSamplesCreationContext<any> {
     constructor(
@@ -12,6 +14,10 @@ export class PropSamplesCreationContext implements IPropSamplesCreationContext<a
         },
     ) {
     }
+
+    getIconList = () => {
+        return getIconList<Icon>(true);
+    };
 
     getCallback = (name: string) => {
         function callbackFn(...args: any[]) {
