+++ uui-docs/src/commonDocs/colorDocs/index.tsx
@@ -1,25 +1,25 @@
 import { IPropDocEditor, TSkin } from '../../types';
-import { colorMapBySkin } from './colorMaps';
+import { COLOR_MAP_BY_SKIN } from './colorMaps';
 import * as React from 'react';
 import { DocBuilder } from '../../DocBuilder';
 import { ColorPicker } from '../../components/colorPicker/ColorPicker';
 
-const getColorDoc = (colorMap?: TColorMap) => {
+const COLOR_DOC_BY_SKIN = {
+    [TSkin.UUI]: getColorDoc(),
+    [TSkin.Electric]: getColorDoc(),
+    [TSkin.Loveship]: getColorDoc(COLOR_MAP_BY_SKIN[TSkin.Loveship]),
+    [TSkin.Promo]: getColorDoc(COLOR_MAP_BY_SKIN[TSkin.Promo]),
+};
+
+export function getColorDocBySkin(skin: TSkin) {
+    return COLOR_DOC_BY_SKIN[skin];
+}
+
+function getColorDoc(colorMap?: TColorMap) {
     return new DocBuilder<{ color: string }>({ name: 'Color' }).prop('color', {
         editorType: getColorPickerComponent(colorMap),
     });
-};
-
-const bySkin = {
-    [TSkin.UUI]: getColorDoc(),
-    [TSkin.Electric]: getColorDoc(),
-    [TSkin.Loveship]: getColorDoc(colorMapBySkin[TSkin.Loveship]),
-    [TSkin.Promo]: getColorDoc(colorMapBySkin[TSkin.Promo]),
-};
-
-export const getColorDocBySkin = (skin: TSkin) => {
-    return bySkin[skin];
-};
+}
 
 type TColorMap = {
     [colorName: string]: string
