+++ uui-docs/src/docBuilderGen/index.ts
@@ -1,1 +1,2 @@
 export { useDocBuilderGen } from './hooks/useDocBuilderGen';
+export * from './docBuilderGenTypes';
