+++ uui-docs/src/index.ts
@@ -1,6 +1,5 @@
 export * from './DocBuilder';
 export * from './types';
-export * from './typesDocConfig';
 export * from './dataSources';
 export * from './demoApi';
 export * from './demoData';
@@ -9,3 +8,4 @@ export * from './commonDocs';
 export * from './peEditors';
 export * from './peContexts';
 export * from './docBuilderGen';
+export * from './docsGen/sharedTypes';
