+++ uui-docs/src/peEditors/iconEditor.tsx
@@ -1,20 +1,12 @@
 import React from 'react';
 import { IconList, IPropDocEditor } from '../types';
 import { IconPickerWithInfo } from '../components/iconPicker/IconPicker';
+import { Icon } from '@epam/uui-core';
 
-/**
- * It's a temporary hack, because currently we don't support collecting all icons in the "uui-docs" module like in "app".
- * The "app" uses webpack-specific API for this. And "uui-docs" is built using Rollup.
- */
-type TIconListGetterHack = { getIconList?: () => IconList<any>[] };
-
-export function IconEditor(props: IPropDocEditor & TIconListGetterHack) {
-    const { value, onValueChange, getIconList } = props;
-    if (!getIconList) {
-        throw new Error('getIconList is mandatory');
-    }
-    const icons = getIconList();
+export function IconEditor(props: IPropDocEditor<IconList<Icon>>) {
+    const { value, onValueChange, examples } = props;
     const editable = { value, onValueChange };
+    const icons = examples.map((e) => e.value);
     return (
         <IconPickerWithInfo icons={ icons } { ...editable } enableInfo={ true } />
     );
