+++ uui-docs/src/types.ts
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import { IDemoApi } from './demoApi';
-
-export * from './sharedTypes';
+import { Icon } from '@epam/uui-core';
 
 export interface DemoComponentProps<TProps = any> {
     DemoComponent: React.ComponentType<TProps> | React.NamedExoticComponent<TProps>;
@@ -26,6 +25,13 @@ export interface IPropSamplesCreationContext<TProps = {}> {
     getSelectedProps(): TProps;
     demoApi: IDemoApi;
     forceUpdate: () => void;
+
+    /**
+     * Currently, the "uui-docs" module is built using Rollup
+     * and therefore can't use webpack-specific API (require.context)
+     * to collect all icons from the epam-assets module. So it's a workaround.     *
+     */
+    getIconList?: () => IconList<Icon>[];
 }
 
 export type PropExampleObject<TProp> = {
