--- uui-docs/src/typesDocConfig.ts
@@ -1,36 +1,0 @@
-import * as React from 'react';
-import { TDocContext, TDocsGenExportedType, TSkin } from './types';
-import { DocBuilder } from './DocBuilder';
-
-export type TDocConfig = {
-    /**
-     * The React component's tag name
-     */
-    name: string;
-    /**
-     * Contexts needed for this doc. The TDocContext.Default is used when no contexts are provided.
-     */
-    contexts?: TDocContext[];
-    bySkin: {
-        [key in TSkin]?: {
-            /**
-             * Identifier of the React component props Typescript type.
-             */
-            type: TDocsGenExportedType,
-            /**
-             * The actual React component
-             */
-            component: React.ComponentType<any> | any,
-            /**
-             * Override doc for this specific skin.
-             * @param doc
-             */
-            doc?: (doc: DocBuilder<any>) => void
-        }
-    };
-    /**
-     * Override doc for all listed skins. Skin-specific override has higher priority.
-     * @param doc
-     */
-    doc?: (doc: DocBuilder<any>) => void;
-};
